refactor: 테스트 상수 클래스로 분리
