+++ src/test/kotlin/blackjack/business/CardFixture.kt
@@ -1,0 +1,10 @@
+package blackjack.business
+
+object CardFixture {
+    val SPACE_ACE = Card(Suit.SPADE, Rank.ACE)
+    val SPACE_EIGHT = Card(Suit.SPADE, Rank.EIGHT)
+    val SPACE_TWO = Card(Suit.SPADE, Rank.TWO)
+    val SPACE_THEN = Card(Suit.SPADE, Rank.TEN)
+    val SPACE_THREE = Card(Suit.SPADE, Rank.THREE)
+    val SPACE_NINE = Card(Suit.SPADE, Rank.NINE)
+}
