+++ src/test/kotlin/blackjack/business/FixDrawConditionStrategy.kt
@@ -1,0 +1,7 @@
+package blackjack.business
+
+class FixDrawConditionStrategy : DrawConditionStrategy {
+    override fun shouldDraw(playerName: String): Boolean {
+        return true
+    }
+}
