+++ src/test/kotlin/blackjack/business/FixPlayerOutputHandler.kt
@@ -1,0 +1,11 @@
+package blackjack.business
+
+import blackjack.view.PlayerOutputHandler
+
+class FixPlayerOutputHandler : PlayerOutputHandler {
+    val printedPlayer: MutableList<Player> = mutableListOf()
+
+    override fun print(player: Player) {
+        printedPlayer.add(player)
+    }
+}
