+++ src/test/kotlin/blackjack/business/FixSelectionStrategy.kt
@@ -1,0 +1,7 @@
+package blackjack.business
+
+class FixSelectionStrategy : CardSelectionStrategy {
+    override fun selectCard(cards: List<Card>): Card {
+        return cards.first()
+    }
+}
