+++ src/test/kotlin/blackjack/business/PlayerTest.kt
@@ -1,5 +1,7 @@
 package blackjack.business
 
+import blackjack.business.CardFixture.SPACE_ACE
+import blackjack.business.CardFixture.SPACE_EIGHT
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
@@ -13,7 +15,7 @@ class PlayerTest {
         val player = Player.from("pobi")
 
         // when
-        player.addCard(PlayerCardDeskTest.SPACE_ACE)
+        player.addCard(SPACE_ACE)
 
         // then
         player.cards.size shouldBe 1
@@ -24,8 +26,8 @@ class PlayerTest {
     fun `플레이어가 추가로 카드를 뽑을 수 있는지 확인한다`() {
         // given
         val player = Player.from("pobi")
-        player.addCard(PlayerCardDeskTest.SPACE_ACE)
-        player.addCard(PlayerCardDeskTest.SPACE_EIGHT)
+        player.addCard(SPACE_ACE)
+        player.addCard(SPACE_EIGHT)
 
         // when
         val canDrawCard = player.canDrawCard()
@@ -38,8 +40,8 @@ class PlayerTest {
     fun `플레이어의 스코어를 확인한다`() {
         // given
         val player = Player.from("pobi")
-        player.addCard(PlayerCardDeskTest.SPACE_ACE)
-        player.addCard(PlayerCardDeskTest.SPACE_EIGHT)
+        player.addCard(SPACE_ACE)
+        player.addCard(SPACE_EIGHT)
 
         // when
         val score = player.score
@@ -51,7 +53,7 @@ class PlayerTest {
     @Test
     fun `플레이어 생성시 PlayerCards 주입`() {
         // given
-        val playerCards = PlayerCards(listOf(PlayerCardDeskTest.SPACE_ACE, PlayerCardDeskTest.SPACE_EIGHT))
+        val playerCards = PlayerCards(listOf(SPACE_ACE, SPACE_EIGHT))
 
         // when
         val player = Player.from("pobi", playerCards)
